requirements.txt
Jinja2==2.8
MarkupSafe==0.23
oauthlib==1.1.2
py-trello==0.6.1
python-dateutil==2.5.3
requests==2.11.1
requests-oauthlib==0.6.2
six==1.10.0
